Add C# textobjects queries for classes, functions and comments (#23)
Support for selection around and inside with the following keybindings

- `vif` Select inside functions
- `vaf` Select around functions
- `vic` Select inside classes
- `vac` Select around classes

This also works with lambdas with one exception. When using the "around
function" motion it will use visual line mode. So if the arguments are
on the same line as around will select the hole block.

Also adds support for navigations with `[f` has support for methods and
interface methods with no body.